# doc/api.yaml
openapi: 3.0.3
info:
  title: WasaPhotoDavidePietragalla
  description: The API for WASAPhoto by Davide Pietragalla
  version: 1.0.0
...
paths:
 /session:
  
    post:
      tags: ["login"]
      summary: It logs in the user
      description: It creates an user if it doesn't exist yet, and it returns the user's identifier
      operationId: doLogin
      requestBody:
        description: User's details
        content:
          application/json:
            schema:
              description: Information to log in the user
              type: object
              properties:
                name:
                  description: Nickname
                  type: string
                  example: "Maria"
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
        required: true
      responses:
        '201':
          description: Log-in successfully executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/full_user"
#_______________________________________________________________
...
 /user/{id}/followers:
  
    parameters:
      - $ref: "#/components/parameters/identifier"
      - $ref: "#/components/parameters/subject_nick"
      
    put:
      tags: ["follow"]
      summary: It allows to follow an user
      description: Used when an user decides to follow another
      operationId: followUser
      responses:
        '204':
          $ref: "#/components/responses/no_content"
        '401':
          $ref: "#/components/responses/unauthorized"
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      security:
        - bearerAuth: []
        
    delete:
      tags: ["follow"]
      summary: It allows to unfollow an user
      description: Used when an user decides to unfollow another
      operationId: unfollowUser
      responses:
        '204':
          $ref: "#/components/responses/no_content"
        '401':
          $ref: "#/components/responses/unauthorized"
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      security:
        - bearerAuth: []
#_______________________________________________________________
 /user/{id}/banned_users:
  
    parameters:
      - $ref: "#/components/parameters/identifier"
      - $ref: "#/components/parameters/subject_nick"
  
    put:
      tags: ["ban"]
      summary: It blocks an user
      description: Used when an user decides to ban another
      operationId: banUser
      responses:
        '204':
          $ref: "#/components/responses/no_content"
        '401':
          $ref: "#/components/responses/unauthorized"
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      security:
        - bearerAuth: []
    
    delete:
      tags: ["ban"]
      summary: It unblocks an user
      description: Used when an user decides to unban another
      operationId: unbanUser
      responses:
        '204':
          $ref: "#/components/responses/no_content"
        '401':
          $ref: "#/components/responses/unauthorized"
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      security:
        - bearerAuth: []
#_______________________________________________________________    
...
 /user/{id}/posts/likes:
  
    parameters: 
        - $ref: "#/components/parameters/identifier"
        - $ref: "#/components/parameters/post_id"
    
    put:
      tags: ["likes"]
      summary: It adds a like
      description: It adds a like to the collection of likes
      operationId: likePhoto
      responses:
        '204':
          $ref: '#/components/responses/no_content'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      security:
        - bearerAuth: []
    
    delete:
      tags: ["likes"]
      summary: It removes a like
      description: It removes a like from the collection of likes
      operationId: unlikePhoto
      responses:
        '204':
          $ref: '#/components/responses/no_content'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      security:
        - bearerAuth: []
#_______________________________________________________________
 /user/{id}/posts/comments:
  
    parameters: 
        - $ref: "#/components/parameters/identifier"
    
    post:
      tags: ["comments"]
      summary: It adds a comment to a post
      description: It adds a comment to a post in a profile
      operationId: commentPhoto
      parameters: 
        - $ref: "#/components/parameters/post_id"
      requestBody:
        description: It contains the content of the comment
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/comment"
        required: true
      responses:
        '201':
          $ref: '#/components/responses/comment_added'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      security:
        - bearerAuth: []
    
    delete:
      tags: ["comments"]
      summary: It deletes a comment from a post
      description: It deletes a comment from a post in a profile
      operationId: uncommentPhoto
      parameters:
        - $ref: "#/components/parameters/comment_id"
      responses:
        '204':
          $ref: '#/components/responses/no_content'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      security:
        - bearerAuth: []
#_______________________________________________________________
components:
  parameters:
...
    identifier:
      name: id
      in: path
      description: The unique id of the user
      required: true
      schema:
        $ref: "#/components/schemas/identifier"
    subject_nick:
      name: subject_nick
      in: query
      description: Nickname of a subject
      required: true
      schema:
        $ref: "#/components/schemas/user_nickname"
    post_id:
      name: post_id
      in: query
      description: A post identifier
      required: true
      schema: 
        $ref: "#/components/schemas/identifier"
    comment_id:
      name: comment_id
      in: query
      description: A comment identifier
      required: true
      schema:
        $ref: "#/components/schemas/identifier"
      example: 123
 #_______________________________________________________________ 
  schemas:
    user_nickname:
      description: Any string between 3-16 characters
      type: string
      minLength: 3
      maxLength: 16
      pattern: '^.*?$'
      example: "Jojo"
    identifier:
      description: Any number between 0-9999
      type: integer
      format: int64
      minimum: 0
      maximum: 9999
      readOnly: true
      example: 965
    full_user:
      description: Contains the *unique identifier and the nickname of a user
      type: object
      properties:
        user_id:
          $ref: "#/components/schemas/identifier"
        nickname:
          $ref: "#/components/schemas/user_nickname"
      example:
        user_id: 966
        nickname: "NeXus"
...
    comment:
      description: A comment containing the text of the comment and details of the owner
      type: object
      properties:
        user_id: 
          $ref: "#/components/schemas/identifier"
        nickname: 
          $ref: "#/components/schemas/user_nickname"
        comment_id:
          $ref: "#/components/schemas/identifier"  
        comment:
          type: string
          minLength: 1
          maxLength: 30
          pattern: '^.*?$'
          example: "Cool brother"
          description: Content of the comment
      example:
        user_id: 888
        nickname: "BadKiddo"
        comment_id: 167
        comment: "WOW WHAT A SCENERY"          
...
  responses:
...
    comment_added:
      description: This response is associated to the 201 http status(Comment has been added successfully and the method returns the comment's id.)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/identifier"
    no_content:
      description: This response is associated to the 204 http status (There is no content to send for this request)
...
    unauthorized:
      description: This response is associated to the 401 http status (The client must authenticate itself to get the requested response)
    forbidden:
      description: This response is associated to the 403 http status (The client does not have access rights to the content)
    not_found:
      description: This response is associated to the 404 http status (The server cannot find the requested resource.)
    internal_server_error:
      description: This response is associated to the 500 http status (The server has encountered a situation it does not know how to handle.)
#_______________________________________________________________
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT